# Speed up Shamir secret sharing split and combine