@Legrandin 2020-06-21
Thanks, I merged it with c10bcbac4e92c0c3a7a393ca96a9b1feb9d3cdcb. Still, I left `coeff_0_l` as it provides a level of blinding for the multiplications involving the secret value.